% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinregClass.R
\docType{class}
\name{Linreg-class}
\alias{Linreg-class}
\alias{LinregClass}
\title{A Reference Class to a linear regression model.}
\description{
A Reference Class to a linear regression model.
}
\section{Fields}{

\describe{
\item{\code{formlua}}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}
}}

