% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Perform linear regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{data}{a data frame or (or object coercible by
\code{\link{as.data.frame})}.}

\item{formlua}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}
}
\value{
\code{linreg} returns an object of class "\code{Linreg}".

  An object of class "\code{Linreg}" have the following methods:
  \item{coefficients()}{a named vector of coefficients}
  \item{residuals()}{the residuals, that is response minus fitted values}
  \item{pred()}{predicted values of of a model}
}
\description{
Use \code{linreg} to perform linear regressions.
}
\examples{
linreg(formula = Sepal.Length ~ Sepal.Width, data = iris)

\dontrun{
linreg(TRUE, TRUE)
}
}
\seealso{
\code{\link{lm}}, \code{\link{class}}, \code{\link{formula}}
}

