% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrlinreg.R
\name{qrlinreg}
\alias{qrlinreg}
\title{Perform linear regression}
\usage{
qrlinreg(x, y, ...)
}
\arguments{
\item{formlua}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}

\item{data}{a data frame or (or object coercible by
\code{\link{as.data.frame}}).}
}
\value{
\code{qrlinreg} returns an
 object of class "\code{Linreg}".

  An object of class "\code{Linreg}" have the following methods:
  \item{coefficients()}{a named vector of coefficients}
  \item{residuals()}{the residuals, that is response minus fitted values}
  \item{pred()}{predicted values of of a model}
}
\description{
Use \code{qrlinreg} to perform linear regressions.
}
\examples{
qrlinreg(formula = Sepal.Length ~ Sepal.Width, data = iris)

\dontrun{
qrlinreg(TRUE, TRUE)
}
}
\seealso{
\code{\link{qr}}, \code{\link{class}}, \code{\link{formula}}
}

