% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{ridgereg function.}
\arguments{
\item{formula}{An object of class formula..}

\item{data}{A data_set.}

\item{lambda}{A choosen penilty}
}
\value{
A model estimation.
}
\description{
ridgereg function.
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{pred()}}{Returns a vector of predicted values}

\item{\code{print()}}{Prints a formula and the coefficients.}
}}

