% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labb4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg function.}
\arguments{
\item{formula}{An object of class formula..
@param data A data_set.
@return A model estimation.}
}
\description{
Linreg function.
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{plot()}}{Plots the residuals against fitted values and Scale-Location.}

\item{\code{pred()}}{Returns a vector of predicted values}

\item{\code{print()}}{Prints a formula and the coefficients.}

\item{\code{resid()}}{Returns a vector of the residuals}

\item{\code{summary()}}{Returns the coefficients, standard error, t-values and p-values}
}}

