% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labb4.R
\name{linreg}
\alias{linreg}
\title{Linreg function.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{An object of class formula..}

\item{data}{A data_set.}
}
\value{
A model estimation.
}
\description{
Linreg function.
}
